requirements.txt
fastapi==0.116.1
mysql-connector-python==9.4.0
pandas==2.3.1
pydantic==2.11.7
pytest== 8.4.1
streamlit==1.48.1
requests==2.32.4
uvicorn==0.35.0




